[FIX] One class - one file
