+++ payment_cielo/models/__init__.py
@@ -1,3 +1,5 @@
 # -*- coding: utf-8 -*-
 
-from . import payment
+from . import payment_acquirer
+from . import payment_transaction
+from . import payment_token
